organization to tools category
I'd call this a tool to extend the basic features of Nextcloud (not like ONLYOFFICE, which adds entirely new things, but a thing that extends the basics). Not customization as it doesn't really change existing behaviour. It isn't communication as it doesn't let users communicate per-se, though it lets the admin communicate to users... Let's just stick with tool?

It doesn't fit, imho, with calendar, tasks, deck or other apps users use to organize and plan.