Support to send announcements to users via email
This PR is based on #140. Thanks to @nickvergessen for the great
groundwork <3

In frontend, We use remark plugin strip-markdown to strip markdown
formatting from the message. Remark was chosen as that's the library
used by juliushaertl/vue-richtext as well.

The options for strip-markdown are changed to keep blockquote and
listItem formatting.

The stripped message is given to the backend as extra parameter
`plainMessage`.

The backend stores plainMessage in the DB and uses it to build the
email body.

Fixes: #65

Signed-off-by: Jonas <jonas@freesources.org>