Update NewForm.vue
Adjusting the groups selection to the top of the popup.

It has been crippled, since its at the bottom of the popup.  You can no longer search groups properly and the popup work ok on the CHROME browser.  So moving it to the top of the popup doesn't hide all the searched groups.

Signed-off-by: spoiledtechie <spoiledtechie@gmail.com>