Update APIController.php
changing 25 to 50.

Searching for these groups doesn't seem to actually work.  Whether its the 25 that breaks this or the search that doesn't work on the group manager, I am unsure.

Placing this limitation of 25 on this, doesn't show more than 25 groups ever.  As the search in the box doesn't make a round trip to the server.  We need this increased to 50 to at least allow 50 groups to be shown here.  

We presently have 36 groups inside of our installation.  I don't believe we will hit 50.

Signed-off-by: spoiledtechie <spoiledtechie@gmail.com>